Fastaextractor removed
