+++ example/models/tal1_model/dataloader.py
@@ -7,7 +7,6 @@ import numpy as np
 import pandas as pd
 import pybedtools
 from pybedtools import BedTool
-from kipoi.helper import FastaExtractor
 from kipoi.data import Dataset
 from kipoi.metadata import GenomicRanges
 import linecache
